status map test

+++ server-setup/exportStateServiceFiles.py
@@ -22,9 +22,12 @@ if __name__ == "__main__":
 
     arcpy.env.overwriteOutput = True
 
+    #set arg here for specific state
+    specific_states = ['sc']
+
     #paths
     scriptPath = os.path.dirname(os.path.realpath(__file__))
-    stateServicesMXD = scriptPath + '/stateServices.mxd'
+    stateServicesMXD = scriptPath + '/stateServices_legacy.mxd'
 
     #main state mxd
     all_mxd = arcpy.mapping.MapDocument(stateServicesMXD)
@@ -35,53 +38,43 @@ if __name__ == "__main__":
         if lyr.isGroupLayer: 
         
             stateAbv = str(lyr).lower()
+
+            if stateAbv in specific_states:
         
-            #set state path
-            statePath = scriptPath + '/State/' + stateAbv + '/'
-            
-            #remove state folder if it exists
-            if os.path.exists(statePath):
-                shutil.rmtree(statePath)
-            os.makedirs(statePath)
-            
-            #loop over single state layers inside of group
-            for subLayer in lyr:
-            
-                layerName = str(subLayer.name)
-                print 'This layer is in a group layer: ',stateAbv ,layerName
-
-                if layerName == 'StreamGrid':
+                #set state path
+                statePath = scriptPath + '/State/' + stateAbv + '/'
                 
-                    #turn off pyramids
-                    arcpy.env.pyramid = "PYRAMIDS 0"
-                    
-                    #set raster compression
-                    #arcpy.env.compression = "CCITT_1D"
-                    
-                    raster = os.path.join(statePath, 'streamgrid.tif')
-                    
-                    print '   Reclassifying stream grid raster...', raster
-                    
-                    # Reclassiffy
-                    #using a tiff allows for 1 bit unsigned pixel depth, reducing file size
-                    arcpy.gp.Reclassify_sa(subLayer, "VALUE", "1 1", raster, "DATA")
-                                    
-                    #print '   Exporting stream grid raster...',raster
-                    #arcpy.CopyRaster_management(in_raster=tempRaster, out_rasterdataset=raster, config_keyword="", background_value="", nodata_value="-128", onebit_to_eightbit="NONE", colormap_to_RGB="NONE", pixel_type="1_BIT", scale_pixel_value="NONE", RGB_to_Colormap="NONE", format="TIFF", transform="NONE")
+                #remove state folder if it exists
+                if os.path.exists(statePath):
+                    shutil.rmtree(statePath)
+                os.makedirs(statePath)
+                
+                #loop over single state layers inside of group
+                for subLayer in lyr:
+                
+                    layerName = str(subLayer.name)
+                    print 'This layer is in a group layer: ',stateAbv ,layerName
 
-                    #delete tempRaster
-                    #arcpy.Delete_management(tempRaster)
+                    if layerName == 'StreamGrid':
                     
-                    #delete pyramids
-                    #if os.path.exists(statePath + 'StreamGrid.rrd'):
-                    #    print '      DELETING raster pyramids....'
-                    #    os.remove(statePath + 'StreamGrid.rrd')
-                    
-                #check for other state layers
-                else:
-                    shape = os.path.join(statePath, layerName.lower() + '.shp')
-                    print '   Exporting other state layer:', shape
-                    arcpy.management.CopyFeatures(subLayer, shape)
+                        #turn off pyramids
+                        arcpy.env.pyramid = "PYRAMIDS 0"
+                        
+                        raster = os.path.join(statePath, 'streamgrid.tif')
+                        
+                        print '   Reclassifying stream grid raster...', raster
+                        
+                        # Reclassiffy
+                        #using a tiff allows for 1 bit unsigned pixel depth, reducing file size
+                        arcpy.gp.Reclassify_sa(subLayer, "VALUE", "1 1", raster, "DATA")
+                                                               
+                    #check for other state layers
+                    else:
+                        shape = os.path.join(statePath, layerName.lower() + '.shp')
+                        print '   Exporting other state layer:', shape
+                        arcpy.management.CopyFeatures(subLayer, shape)
+            else:
+                'skipping this state:',stateAbv
 
             #paise the loop
             #raw_input("Press Enter to continue...")
